requirements.txt
bcrypt==5.0.0
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
DBUtils==3.1.2
dnspython==2.8.0
eventlet==0.40.4
Flask==3.1.2
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SocketIO==5.6.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.0
gunicorn==21.2.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
msgspec==0.20.0
packaging==26.0
pillow==12.1.0
PyJWT==2.10.1
PyMySQL==1.1.2
python-docx==1.2.0
python-dotenv==1.0.0
python-engineio==4.13.0
python-socketio==5.16.0
requests==2.32.5
simple-websocket==1.1.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
wsproto==1.3.2
WTForms==3.2.1
